requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
Pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
openai==1.58.1
langchain==0.3.13
langchain-openai==0.2.14
googlemaps==4.10.0
celery==5.4.0
redis==5.2.1
drf-yasg==1.21.8
gunicorn==23.0.0
whitenoise==6.8.2
